Move cache utilities to lib/
