+++ packages/dictionaries/src/lib/index.ts
@@ -1,4 +1,8 @@
+export { default as Cache } from './Cache';
+export { default as DiskCache } from './DiskCache';
 export { default as ensureDirectoryExists } from './ensureDirectoryExists';
 export { default as getDictionaryFilepath } from './getDictionaryFilepath';
+export { default as LayeredCache } from './LayeredCache';
+export { default as MemoryCache } from './MemoryCache';
 export { default as readFile } from './readFile';
 export { default as writeFile } from './writeFile';
